Mike 9/07/2018 v1.2.1.5
EDIT ACCOUNT: I add the functionality of change password, and delete account, also i change the button of place. I modified some links to open the correct login when you log in in the application. RAMON and BEN, please download this changes first before you upload your changes.
i need to modify the code, to add the new condition to the user, that the user can log in, if the user is active.