Mike 03/05/2018 - 1.2.1.13
Edit profile: I change the functionality of edit profile in the sense that you need to check the availability of email and the username, this is a temporary functionality, I can not resolve the problem of the promises.

We need to add design in the buttons and the previous image of the profile.

All the functionality is ready in Edit Profile.

PLEASE RAMOM AND BEN check the updates in the GITHUB before you upload your changes.  Danke!!!!