Mike 31/05/2018 v1.2.1.15
The screen from the menu board that is almost complete but it is missing for example implement the galery to the user can change their image in the board.

I charged the image with the original image because the app take a little time (2 seconds or more) to create the thumbnail and thats enough to not charge the thumbnail,  then i charged the last image that the user upload to display the image of the service.